refactor: Move component files by its purpose
- add common directory for commonly used component
- add blog directory for blog service level component
- update tailwind config to apply css
- update gatsby configs
- update page components
