+++ gatsby-browser.tsx
@@ -18,7 +18,7 @@ import { RecoilRoot } from "recoil";
  */
 import "./src/styles/global.css";
 import "./src/styles/syntax-highlight.css";
-import Layout from "./src/components/Layout";
+import Layout from "./src/components/blog/Layout";
 
 /**
  * Type modules
