+++ src/components/blog/index.ts
@@ -1,0 +1,4 @@
+export { Category } from "./Category";
+export { PostCard } from "./PostCard";
+export { SEO } from "./SEO";
+export { TableOfContents } from "./TableOfContents";
