--- src/components/index.ts
@@ -1,9 +1,0 @@
-export { Button } from "./Button";
-export { Category } from "./Category";
-export { Footer } from "./Footer";
-export { Header } from "./Header";
-export { ImageFallback } from "./ImageFallback";
-export { PostCard } from "./PostCard";
-export { Progress } from "./Progress";
-export { SEO } from "./SEO";
-export { TableOfContents } from "./TableOfContents";
