+++ src/pages/500.tsx
@@ -6,8 +6,8 @@ import { TbPray } from "react-icons/tb";
 /**
  * Internal modules
  */
-import { Button } from "../components/Button";
-import { SEO } from "../components/SEO";
+import { Button } from "../components/common";
+import { SEO } from "../components/blog";
 import { useSiteMetadata } from "../hooks/useSiteMetadata";
 
 /**
