+++ tailwind.config.js
@@ -1,7 +1,7 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
     content: [
-        "./src/components/*.{js,jsx,ts,tsx}",
+        "./src/components/**/*.{js,jsx,ts,tsx}",
         "./src/templates/*.{js,jsx,ts,tsx}",
         "./src/pages/*.{js,jsx,ts,tsx}",
         "./content/**/*.mdx"
